feat: 1. Подключение к базе в веб приложении. Хранение вакансий. [#504859]
- Добавить зависимости в pom.xml на JDBC драйвер PostgreSQL, реализацию пула соединений commons-dbcp2 и клиент для работы с БД Sql2O;

- создать БД и создать таблицы, которым соответствуют наши доменные модели;

- Внесите строки для подключения к БД в application.properties;

- Создать класс конфигурацию ru.job4j.dreamjob.configuration.DatasourceConfiguration

- Реализовать работу с БД: реализация Sql2oCityRepository, Sql2oFileRepository, Sql2oVacancyRepository

- изменить Vacancy: написать маппинг в виде мапы COLUMN_MAPPING, ключи которой это столбцы из БД, а значения названия полей;

- внедрить в классах сервисов реализацию интерфейсов репозиториев, работающих с БД(достаточно изменить названия параметров в конструкторах)

+++ src/main/java/ru/job4j/dreamjob/service/SimpleCandidateService.java
@@ -39,7 +39,13 @@ public class SimpleCandidateService implements CandidateService {
 
     @Override
     public boolean deleteById(int id) {
-        return candidateRepository.deleteById(id);
+        boolean isDeleted = false;
+        var fileOptional = findById(id);
+        if (fileOptional.isPresent()) {
+            isDeleted = candidateRepository.deleteById(id);
+            fileService.deleteById(fileOptional.get().getFileId());
+        }
+        return isDeleted;
     }
 
     @Override
